Move types inline
