+++ lib/adapters/adapter.ts
@@ -1,17 +1,10 @@
-// Public types for Figbird consumers
-
 /**
  * Generic response wrapper for all operations
  */
-export interface Response<TData, TMeta = Record<string, unknown>> {
+export interface QueryResponse<TData, TMeta = Record<string, unknown>> {
   data: TData
   meta: TMeta
 }
-
-/**
- * Event types supported by Figbird
- */
-export type EventType = 'created' | 'updated' | 'patched' | 'removed'
 
 /**
  * Event handlers for real-time updates
@@ -28,9 +21,9 @@ export interface EventHandlers<T> {
  */
 export interface Adapter<T = unknown, TParams = unknown> {
   // Required core methods
-  get(serviceName: string, resourceId: string | number, params?: TParams): Promise<Response<T>>
-  find(serviceName: string, params?: TParams): Promise<Response<T[]>>
-  findAll(serviceName: string, params?: TParams): Promise<Response<T[]>>
+  get(serviceName: string, resourceId: string | number, params?: TParams): Promise<QueryResponse<T>>
+  find(serviceName: string, params?: TParams): Promise<QueryResponse<T[]>>
+  findAll(serviceName: string, params?: TParams): Promise<QueryResponse<T[]>>
   mutate(serviceName: string, method: string, args: unknown[]): Promise<T>
 
   // Optional real-time support
